--- vue-self-master/README.md
@@ -1,1 +1,0 @@
-#zeji_pro
