--- vue-self-master/npm-debug.log
@@ -1,50 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ 'C:\\Users\\tempus\\node.exe',
-1 verbose cli   'C:\\Users\\tempus\\node_modules\\npm\\bin\\npm-cli.js',
-1 verbose cli   'run',
-1 verbose cli   'dev',
-1 verbose cli   '--scripts-prepend-node-path=auto' ]
-2 info using npm@3.10.10
-3 info using node@v6.11.1
-4 verbose run-script [ 'predev', 'dev', 'postdev' ]
-5 info lifecycle cli@1.0.0~predev: cli@1.0.0
-6 silly lifecycle cli@1.0.0~predev: no script for predev, continuing
-7 info lifecycle cli@1.0.0~dev: cli@1.0.0
-8 verbose lifecycle cli@1.0.0~dev: unsafe-perm in lifecycle true
-9 verbose lifecycle cli@1.0.0~dev: PATH: C:\Users\tempus\node_modules\npm\bin\node-gyp-bin;D:\vue-self-master\node_modules\.bin;C:\Windows\system32;C:\Windows;C:\Windows\System32\Wbem;C:\Windows\System32\WindowsPowerShell\v1.0\;F:\Git\cmd;C:\Users\tempus\;C:\Ruby23-x64\bin;C:\Users\tempus\AppData\Roaming\npm
-10 verbose lifecycle cli@1.0.0~dev: CWD: D:\vue-self-master
-11 silly lifecycle cli@1.0.0~dev: Args: [ '/d /s /c',
-11 silly lifecycle   'webpack-dev-server --hot --config build/webpack.dev.config.js' ]
-12 silly lifecycle cli@1.0.0~dev: Returned: code: 1  signal: null
-13 info lifecycle cli@1.0.0~dev: Failed to exec dev script
-14 verbose stack Error: cli@1.0.0 dev: `webpack-dev-server --hot --config build/webpack.dev.config.js`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (C:\Users\tempus\node_modules\npm\lib\utils\lifecycle.js:255:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (C:\Users\tempus\node_modules\npm\lib\utils\spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:891:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
-15 verbose pkgid cli@1.0.0
-16 verbose cwd D:\vue-self-master
-17 error Windows_NT 6.1.7600
-18 error argv "C:\\Users\\tempus\\node.exe" "C:\\Users\\tempus\\node_modules\\npm\\bin\\npm-cli.js" "run" "dev" "--scripts-prepend-node-path=auto"
-19 error node v6.11.1
-20 error npm  v3.10.10
-21 error code ELIFECYCLE
-22 error cli@1.0.0 dev: `webpack-dev-server --hot --config build/webpack.dev.config.js`
-22 error Exit status 1
-23 error Failed at the cli@1.0.0 dev script 'webpack-dev-server --hot --config build/webpack.dev.config.js'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the cli package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     webpack-dev-server --hot --config build/webpack.dev.config.js
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs cli
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls cli
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
