--- vue-self-master/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "cli",
-  "version": "1.0.0",
-  "description": "项目名称",
-  "main": "index.js",
-  "scripts": {
-    "dev": "webpack-dev-server --hot --config build/webpack.dev.config.js",
-    "build": "webpack --config build/webpack.pro.config.js"
-  },
-  "keywords": [
-    "描述1",
-    "描述2"
-  ],
-  "author": "yuanbin.xie",
-  "license": "ISC",
-  "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-component": "^0.9.1",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "clean-webpack-plugin": "^0.1.16",
-    "css-loader": "^0.26.4",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.10.1",
-    "html-webpack-plugin": "^2.28.0",
-    "style-loader": "^0.13.2",
-    "url-loader": "^0.5.8",
-    "vue-loader": "^11.3.4",
-    "vue-style-loader": "^2.0.5",
-    "vue-template-compiler": "^2.3.3",
-    "webpack": "^2.6.1",
-    "webpack-dev-server": "^2.4.5",
-    "webpack-merge": "^4.1.0"
-  },
-  "dependencies": {
-    "axios": "^0.16.2",
-    "vue": "^2.3.3",
-    "vue-router": "^2.5.3",
-    "vuex": "^2.3.1"
-  }
-}
