--- vue-self-master/src/vuex/getters.js
@@ -1,6 +1,0 @@
-/**
- * Created by Administrator on 2017/3/15.
- */
-export default {
-    //getters
-}
