--- vue-self-master/src/vuex/index.js
@@ -1,20 +1,0 @@
-/**
- * Created by xieyuanbin on 2017/8/8.
- * Email:yuanbin.xie@tempus.cn
- */
-
-import Vue from 'vue'
-import Vuex from 'vuex'
-import state from './state'
-import mutations from './mutations'
-import getters from './getters'
-import actions from './actions'
-
-Vue.use(Vuex)
-
-export default new Vuex.Store({
-    state,
-    mutations,
-    getters,
-    actions
-})
